api/requirements.txt
fastapi==0.75.2
numpy==1.22.2
pandas==1.4.1
psycopg2==2.9.3
psycopg2_binary==2.9.3
pydantic==1.9.0
PyYAML==6.0
scikit_learn==1.0.2
uvicorn==0.15.0
sentence_transformers
pyarrow